extracted CollectionItemRow from CollectionItem and CollectionRow from Collection
