+++ packages/bruno-app/src/components/Sidebar/Collections/Collection/CollectionItem/index.js
@@ -1,871 +1,54 @@
-import React, { useState, useRef, useEffect, useMemo } from 'react';
+import React from 'react';
 import range from 'lodash/range';
 import filter from 'lodash/filter';
-import classnames from 'classnames';
-import { useDrag, useDrop } from 'react-dnd';
-import { getEmptyImage } from 'react-dnd-html5-backend';
-import {
-  IconChevronRight,
-  IconDots,
-  IconFilePlus,
-  IconFolderPlus,
-  IconPlayerPlay,
-  IconEdit,
-  IconCopy,
-  IconClipboard,
-  IconCode,
-  IconFolder,
-  IconTrash,
-  IconSettings,
-  IconInfoCircle,
-  IconTerminal2,
-  IconAppWindow,
-  IconEyeOff
-} from '@tabler/icons';
 import { useSelector, useDispatch } from 'react-redux';
-import { addTab, focusTab, makeTabPermanent } from 'providers/ReduxStore/slices/tabs';
-import { handleMultipleCollectionItemsDrop, sendRequest, showInFolder, pasteItem, saveRequest, cloneItem } from 'providers/ReduxStore/slices/collections/actions';
-import { sanitizeName } from 'utils/common/regex';
-import { formatIpcError } from 'utils/common/error';
-import { toggleCollectionItem, addResponseExample } from 'providers/ReduxStore/slices/collections';
-import { uuid } from 'utils/common';
-import { copyRequest, setFocusedSidebarPath, insertTaskIntoQueue } from 'providers/ReduxStore/slices/app';
-import NewRequest from 'components/Sidebar/NewRequest';
-import NewFolder from 'components/Sidebar/NewFolder';
-import NewApp from 'components/Sidebar/NewApp';
-import RenameCollectionItem from './RenameCollectionItem';
-import CloneCollectionItem from './CloneCollectionItem';
-import DeleteCollectionItems from './DeleteCollectionItems';
-import IgnoreCollectionItem from './IgnoreCollectionItem';
-import RunCollectionItem from './RunCollectionItem';
-import GenerateCodeItem from './GenerateCodeItem';
-import { isItemARequest, isItemAFolder, scrollToTheActiveTab } from 'utils/tabs';
-import { doesRequestMatchSearchText, doesFolderHaveItemsMatchSearchText } from 'utils/collections/search';
-import { getDefaultRequestPaneTab, getItemTypeLabel } from 'utils/collections';
-import toast from 'react-hot-toast';
-import StyledWrapper from './StyledWrapper';
-import NetworkError from 'components/ResponsePane/NetworkError/index';
-import CollectionItemInfo from './CollectionItemInfo/index';
-import CollectionItemIcon from './CollectionItemIcon';
-import ExampleItem from './ExampleItem';
-import ExampleIcon from 'components/Icons/ExampleIcon';
-import {
-  getTabUidForItem as getTabUidForItemSelector,
-  isTabForItemActive as isTabForItemActiveSelector,
-  isTabForItemPresent as isTabForItemPresentSelector
-} from 'src/selectors/tab';
-import { isEqual } from 'lodash';
-import { createEmptyStateMenuItems } from 'utils/collections/emptyStateRequest';
-import {
-  canCollectionItemBeDropped,
-  determineCollectionItemDrop,
-  getInitialExampleName,
-  findParentItemInCollection,
-  getSelectionInfo,
-  getSortedDraggedItems
-} from 'utils/collections/index';
+import { isItemAFolder, isItemARequest } from 'utils/tabs';
 import { sortByNameThenSequence } from 'utils/common/index';
-import { getRevealInFolderLabel } from 'utils/common/platform';
-import CreateExampleModal from 'components/ResponseExample/CreateExampleModal';
-import { openDevtoolsAndSwitchToTerminal } from 'utils/terminal';
-import ActionIcon from 'ui/ActionIcon';
-import MenuDropdown from 'ui/MenuDropdown';
+import { createEmptyStateMenuItems } from 'utils/collections/emptyStateRequest';
 import { useSidebarAccordion } from 'components/Sidebar/SidebarAccordionContext';
-import useKeybinding from 'hooks/useKeybinding';
-import useSidebarSelectionClick from 'hooks/useSidebarSelectionClick';
-import useMultiSelectDragDisabled from 'hooks/useMultiSelectDragDisabled';
-import { clearSidebarSelection } from 'providers/ReduxStore/slices/collections/index';
+import MenuDropdown from 'ui/MenuDropdown';
+import CollectionItemRow from './CollectionItemRow';
 
+/**
+ * Thin recursive wrapper around CollectionItemRow. The row renders the item itself (name,
+ * chevron, menu, drag/drop, multi-select, examples); this wrapper computes the item's grouped
+ * children and, when the folder is expanded, renders them recursively as the row's `children`.
+ */
 const CollectionItem = ({ item, collectionUid, collectionPathname, searchText, openBulkMenu }) => {
-  const { dropdownContainerRef } = useSidebarAccordion();
-  const selectorInput = {
-    itemUid: item.uid,
-    itemPathname: item.pathname,
-    collectionUid
-  };
-
-  const _isTabForItemActiveSelector = isTabForItemActiveSelector(selectorInput);
-  const isTabForItemActive = useSelector(_isTabForItemActiveSelector, isEqual);
-
-  const _isTabForItemPresentSelector = isTabForItemPresentSelector(selectorInput);
-  const isTabForItemPresent = useSelector(_isTabForItemPresentSelector, isEqual);
-
-  const _tabUidForItemSelector = getTabUidForItemSelector(selectorInput);
-  const tabUidForItem = useSelector(_tabUidForItemSelector, isEqual);
-
-  const isSidebarDragging = useSelector((state) => state.app.isDragging);
+  const dispatch = useDispatch();
   const allCollections = useSelector((state) => state.collections.collections);
   const collection = allCollections?.find((c) => c.uid === collectionUid);
-  const { hasCopiedItems } = useSelector((state) => state.app.clipboard);
-  const selectedSidebarUids = useSelector((state) => state.collections.selectedSidebarUids);
-  const isSelected = selectedSidebarUids.includes(item.uid);
-  const isMultiSelected = isSelected && selectedSidebarUids.length > 1;
-  const handleSelectionClick = useSidebarSelectionClick({ uid: item.uid, searchText });
-  const workspaces = useSelector((state) => state.workspaces.workspaces);
-  const activeWorkspaceUid = useSelector((state) => state.workspaces.activeWorkspaceUid);
-  const activeWorkspace = workspaces?.find((w) => w.uid === activeWorkspaceUid);
-  const collectionSortOrder = useSelector((state) => state.collections.collectionSortOrder);
-  const dispatch = useDispatch();
+  const { dropdownContainerRef } = useSidebarAccordion();
 
-  // When dragging a multi-selected row, carry all effectively-selected folders/requests
-  // (excluding collections) so dropping one moves the entire selection together.
-  const multiDragItems = useMemo(() => {
-    if (!isSelected || selectedSidebarUids.length < 2) return null;
-    const { effectiveSelection, hasCollection } = getSelectionInfo({ collections: allCollections, selectedUids: selectedSidebarUids });
-    if (hasCollection) return null;
-    return effectiveSelection.map((entry) => ({ ...entry.item, sourceCollectionUid: entry.collectionUid }));
-  }, [isSelected, selectedSidebarUids, allCollections]);
-
-  const isDragDisabled = useMultiSelectDragDisabled({ isSelected, selectedSidebarUids, allCollections });
-
-  // We use a single ref for drag and drop.
-  const ref = useRef(null);
-  const menuDropdownRef = useRef(null);
-
-  const [renameItemModalOpen, setRenameItemModalOpen] = useState(false);
-  const [deleteItemModalOpen, setDeleteItemModalOpen] = useState(false);
-  const [ignoreItemModalOpen, setIgnoreItemModalOpen] = useState(false);
-  const [createExampleModalOpen, setCreateExampleModalOpen] = useState(false);
-  const [generateCodeItemModalOpen, setGenerateCodeItemModalOpen] = useState(false);
-  const [newRequestModalOpen, setNewRequestModalOpen] = useState(false);
-  const [newFolderModalOpen, setNewFolderModalOpen] = useState(false);
-  const [newAppModalOpen, setNewAppModalOpen] = useState(false);
-  const [runCollectionModalOpen, setRunCollectionModalOpen] = useState(false);
-  const [itemInfoModalOpen, setItemInfoModalOpen] = useState(false);
-  const [examplesExpanded, setExamplesExpanded] = useState(false);
-  const [isKeyboardFocused, setIsKeyboardFocused] = useState(false);
   const hasSearchText = searchText && searchText?.trim()?.length;
   const itemIsCollapsed = hasSearchText ? false : item.collapsed;
   const isFolder = isItemAFolder(item);
 
-  const isCloneable = isFolder || isItemARequest(item) || item.type === 'app';
-
-  // Check if request has examples (only for HTTP requests)
-  const hasExamples = isItemARequest(item) && item.type === 'http-request' && item.examples && item.examples.length > 0;
-
-  // Sidebar shortcuts — only active when this sidebar item has keyboard focus
-  useKeybinding('cloneItem', () => {
-    handleCloneItem();
-    return false;
-  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
-
-  useKeybinding('copyItem', () => {
-    handleCopyItem();
-    return false;
-  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
-
-  useKeybinding('pasteItem', () => {
-    handlePasteItem();
-    return false;
-  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
-
-  useKeybinding('renameItem', () => {
-    setRenameItemModalOpen(true);
-    return false;
-  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
-
-  useKeybinding('newRequest', () => {
-    if (!isFolder) return false;
-    setNewRequestModalOpen(true);
-    return false;
-  }, { enabled: isKeyboardFocused && isFolder, deps: [isKeyboardFocused, isFolder] });
-
-  const [dropType, setDropType] = useState(null); // 'above', 'inside' or 'below'
-
-  const [{ isDragging }, drag, dragPreview] = useDrag({
-    type: isDragDisabled ? 'disabled-drag' : 'collection-item',
-    item: {
-      ...item,
-      sourceCollectionUid: collectionUid,
-      wasSelected: isSelected,
-      ...(multiDragItems ? { multiSelectedItems: multiDragItems } : {})
-    },
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging()
-    }),
-    options: {
-      dropEffect: 'move'
-    }
-  });
-
-  // Auto-scroll to show this item when its tab becomes active
-  useEffect(() => {
-    if (isTabForItemActive && ref.current) {
-      try {
-        ref.current.scrollIntoView({ behavior: 'smooth', block: 'nearest' });
-      } catch (err) {
-        // ignore scroll errors (some environments may not support smooth scrolling)
-      }
-    }
-  }, [isTabForItemActive]);
-
-  const resolveDropFromMonitor = (monitor) => {
-    return determineCollectionItemDrop({
-      item,
-      hoverBoundingRect: ref.current?.getBoundingClientRect(),
-      clientOffset: monitor.getClientOffset()
-    });
-  };
-
-  const canItemBeDropped = ({ draggedItem, targetItem, dropType }) => {
-    return canCollectionItemBeDropped({
-      draggedItem,
-      targetItem,
-      dropType,
-      collectionUid,
-      collectionPathname
-    });
-  };
-
-  const canAnyItemBeDropped = ({ draggedItem, targetItem, dropType }) => {
-    const items = draggedItem.multiSelectedItems?.length > 0 ? draggedItem.multiSelectedItems : [draggedItem];
-    return items.some((i) => canItemBeDropped({ draggedItem: i, targetItem, dropType }));
-  };
-
-  const [{ isOver, canDrop }, drop] = useDrop({
-    accept: 'collection-item',
-    hover: (draggedItem, monitor) => {
-      const { uid: targetItemUid } = item;
-      const { uid: draggedItemUid } = draggedItem;
-
-      if (draggedItemUid === targetItemUid) return;
-
-      const dropType = resolveDropFromMonitor(monitor);
-      if (!dropType) {
-        setDropType(null);
-        return;
-      }
-
-      const _canItemBeDropped = canAnyItemBeDropped({ draggedItem, targetItem: item, dropType });
-
-      setDropType(_canItemBeDropped ? dropType : null);
-    },
-    drop: async (draggedItem, monitor) => {
-      const { uid: targetItemUid } = item;
-      const { uid: draggedItemUid } = draggedItem;
-
-      if (draggedItemUid === targetItemUid) return;
-
-      const dropType = resolveDropFromMonitor(monitor);
-      if (!dropType) return;
-
-      if (!canAnyItemBeDropped({ draggedItem, targetItem: item, dropType })) return;
-
-      const draggedItems = getSortedDraggedItems({
-        draggedItem,
-        allCollections,
-        workspaces,
-        activeWorkspace,
-        collectionSortOrder,
-        searchText
-      });
-
-      // Filter out items that can't be dropped on this target
-      const validDraggedItems = draggedItems.filter((dragged) => {
-        if (dragged.uid === targetItemUid) return false;
-        return canItemBeDropped({ draggedItem: dragged, targetItem: item, dropType });
-      });
-
-      if (validDraggedItems.length > 0) {
-        await dispatch(handleMultipleCollectionItemsDrop({
-          targetItem: item,
-          draggedItems: validDraggedItems,
-          dropType,
-          collectionUid
-        }));
-      }
-
-      if (draggedItem.wasSelected) {
-        dispatch(clearSidebarSelection());
-      }
-
-      setDropType(null);
-    },
-    canDrop: (draggedItem, monitor) => {
-      if (draggedItem.uid === item.uid) return false;
-
-      const dropType = resolveDropFromMonitor(monitor);
-      if (!dropType) return false;
-
-      return canAnyItemBeDropped({ draggedItem, targetItem: item, dropType });
-    },
-    collect: (monitor) => ({
-      isOver: monitor.isOver(),
-      canDrop: monitor.canDrop()
-    })
-  });
-
-  drag(drop(ref));
-  dragPreview(getEmptyImage(), { captureDraggingState: true });
-
-  useEffect(() => {
-    if (!isOver) {
-      setDropType(null);
-    }
-  }, [isOver]);
-
-  const iconClassName = classnames({
-    'rotate-90': !itemIsCollapsed
-  });
-
-  const examplesIconClassName = classnames({
-    'rotate-90': examplesExpanded
-  });
-
-  const itemRowClassName = classnames('flex collection-item-name relative items-center', {
-    'item-focused-in-tab': isTabForItemActive,
-    'item-hovered': isOver && canDrop,
-    'drop-target': isOver && canDrop && dropType === 'inside',
-    'drop-target-above': isOver && canDrop && dropType === 'above',
-    'drop-target-below': isOver && canDrop && dropType === 'below',
-    'item-keyboard-focused': isKeyboardFocused,
-    'collection-item-selected': isSelected,
-    'drag-disabled': isDragDisabled
-  });
-
-  const handleRun = async () => {
-    dispatch(sendRequest(item, collectionUid)).catch((err) =>
-      toast.custom((t) => <NetworkError onClose={() => toast.dismiss(t.id)} />, {
-        duration: 5000
-      })
-    );
-  };
-
-  const handleClick = (event) => {
-    if (handleSelectionClick(event)) return;
-    if (event && event.detail != 1) return;
-    // scroll to the active tab
-    setTimeout(scrollToTheActiveTab, 50);
-    const isRequest = isItemARequest(item);
-    const isApp = item.type === 'app';
-    if (isRequest || isApp) {
-      if (isTabForItemPresent) {
-        dispatch(
-          focusTab({
-            uid: tabUidForItem || item.uid
-          })
-        );
-        return;
-      }
-      dispatch(
-        addTab({
-          uid: item.uid,
-          collectionUid: collectionUid,
-          ...(isRequest ? { requestPaneTab: getDefaultRequestPaneTab(item) } : {}),
-          type: item.type,
-          pathname: item.pathname
-        })
-      );
-    } else {
-      dispatch(
-        addTab({
-          uid: item.uid,
-          collectionUid: collectionUid,
-          type: 'folder-settings',
-          pathname: item.pathname
-        })
-      );
-      if (item.collapsed) {
-        dispatch(
-          toggleCollectionItem({
-            itemUid: item.uid,
-            collectionUid: collectionUid
-          })
-        );
-      }
-    }
-  };
-
-  const handleFolderCollapse = (e) => {
-    e.stopPropagation();
-    e.preventDefault();
-    dispatch(
-      toggleCollectionItem({
-        itemUid: item.uid,
-        collectionUid: collectionUid
-      })
-    );
-  };
-
-  // prevent the parent's double-click handler from firing
-  const handleFolderDoubleClick = (e) => {
-    e.stopPropagation();
-    e.preventDefault();
-  };
-
-  const handleExamplesCollapse = (e) => {
-    e.stopPropagation();
-    e.preventDefault();
-    setExamplesExpanded(!examplesExpanded);
-  };
-
-  // prevent the parent's double-click handler from firing
-  const handleExamplesDoubleClick = (e) => {
-    e.stopPropagation();
-    e.preventDefault();
-  };
-
-  // Handle right-click context menu
-  const handleContextMenu = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    if (isMultiSelected) {
-      openBulkMenu(e);
-      return;
-    }
-
-    menuDropdownRef.current?.show();
-  };
-
-  const indents = range(item.depth);
-
-  // Build menu items for MenuDropdown
-  const buildMenuItems = () => {
-    const items = [];
-
-    if (isFolder) {
-      items.push(
-        {
-          id: 'new-request',
-          leftSection: IconFilePlus,
-          label: 'New Request',
-          onClick: () => setNewRequestModalOpen(true)
-        },
-        {
-          id: 'new-folder',
-          leftSection: IconFolderPlus,
-          label: 'New Folder',
-          onClick: () => setNewFolderModalOpen(true)
-        },
-        {
-          id: 'new-app',
-          leftSection: IconAppWindow,
-          label: 'New App',
-          onClick: () => setNewAppModalOpen(true)
-        },
-        {
-          id: 'run',
-          leftSection: IconPlayerPlay,
-          label: 'Run',
-          onClick: () => setRunCollectionModalOpen(true)
-        }
-      );
-    }
-
-    if (isCloneable) {
-      items.push({
-        id: 'clone',
-        leftSection: IconCopy,
-        label: 'Clone',
-        onClick: handleCloneItem
-      });
-    }
-
-    items.push({
-      id: 'copy',
-      leftSection: IconCopy,
-      label: 'Copy',
-      onClick: handleCopyItem
-    });
-
-    if (isFolder && hasCopiedItems) {
-      items.push({
-        id: 'paste',
-        leftSection: IconClipboard,
-        label: 'Paste',
-        onClick: handlePasteItem
-      });
-    }
-
-    items.push(
-      {
-        id: 'rename',
-        leftSection: IconEdit,
-        label: 'Rename',
-        onClick: () => setRenameItemModalOpen(true)
-      }
-    );
-    if (!isFolder && isItemARequest(item) && !(item.type === 'http-request' || item.type === 'graphql-request')) {
-      items.push({
-        id: 'run',
-        leftSection: IconPlayerPlay,
-        label: 'Run',
-        onClick: () => {
-          handleRun();
-        }
-      });
-    }
-
-    if (!isFolder && (item.type === 'http-request' || item.type === 'graphql-request')) {
-      items.push({
-        id: 'generate-code',
-        leftSection: IconCode,
-        label: 'Generate Code',
-        onClick: handleGenerateCode
-      });
-    }
-
-    if (!isFolder && isItemARequest(item) && item.type === 'http-request') {
-      items.push({
-        id: 'create-example',
-        leftSection: ExampleIcon,
-        label: 'Create Example',
-        onClick: () => setCreateExampleModalOpen(true)
-      });
-    }
-
-    items.push(
-      {
-        id: 'show-in-folder',
-        leftSection: IconFolder,
-        label: getRevealInFolderLabel(),
-        onClick: handleShowInFolder
-      }
-    );
-
-    if (isFolder) {
-      items.push({
-        id: 'ignore',
-        leftSection: IconEyeOff,
-        label: 'Ignore',
-        onClick: () => setIgnoreItemModalOpen(true)
-      });
-    }
-
-    items.push({ id: 'separator-1', type: 'divider' });
-
-    items.push({
-      id: 'info',
-      leftSection: IconInfoCircle,
-      label: 'Info',
-      onClick: () => setItemInfoModalOpen(true)
-    });
-
-    if (isFolder) {
-      items.push(
-        {
-          id: 'settings',
-          leftSection: IconSettings,
-          label: 'Settings',
-          onClick: viewFolderSettings
-        },
-        {
-          id: 'open-terminal',
-          leftSection: IconTerminal2,
-          label: 'Open in Terminal',
-          onClick: async () => {
-            const folderCwd = item.pathname || collectionPathname;
-            await openDevtoolsAndSwitchToTerminal(dispatch, folderCwd);
-          }
-        }
-      );
-    }
-
-    items.push({
-      id: 'delete',
-      leftSection: IconTrash,
-      label: 'Delete',
-      className: 'delete-item',
-      onClick: () => setDeleteItemModalOpen(true)
-    });
-
-    return items;
-  };
-
-  const className = classnames('flex flex-col w-full', {
-    'is-sidebar-dragging': isSidebarDragging
-  });
-
-  if (searchText && searchText.length) {
-    if (isItemARequest(item)) {
-      if (!doesRequestMatchSearchText(item, searchText)) {
-        return null;
-      }
-    } else {
-      if (!doesFolderHaveItemsMatchSearchText(item, searchText)) {
-        return null;
-      }
-    }
-  }
-
-  const handleDoubleClick = (event) => {
-    dispatch(makeTabPermanent({ uid: tabUidForItem || item.uid }));
-  };
-
-  // Sort items by their "seq" property.
-  const sortItemsBySequence = (items = []) => {
-    return items.sort((a, b) => a.seq - b.seq);
-  };
-
-  const handleShowInFolder = () => {
-    dispatch(showInFolder(item.pathname)).catch((error) => {
-      console.error('Error opening the folder', error);
-      toast.error('Error opening the folder');
-    });
-  };
-
-  const handleCreateExample = async (name, description = '') => {
-    const exampleData = {
-      name: name,
-      description: description,
-      status: 200,
-      statusText: 'OK',
-      headers: [],
-      body: {
-        type: 'text',
-        content: ''
-      }
-    };
-
-    // Calculate the index where the example will be saved
-    const existingExamples = item.draft?.examples || item.examples || [];
-    const exampleIndex = existingExamples.length;
-    const exampleUid = uuid();
-
-    dispatch(addResponseExample({
-      itemUid: item.uid,
-      collectionUid: collectionUid,
-      example: {
-        ...exampleData,
-        uid: exampleUid
-      }
-    }));
-
-    // Save the request
-    await dispatch(saveRequest(item.uid, collectionUid, true));
-
-    // Task middleware will track this and open the example in a new tab once the file is reloaded
-    dispatch(insertTaskIntoQueue({
-      uid: exampleUid,
-      type: 'OPEN_EXAMPLE',
-      collectionUid: collectionUid,
-      itemUid: item.uid,
-      exampleIndex: exampleIndex,
-      // Freshly created examples start blank, so open the tab in edit mode.
-      openInEditMode: true
-    }));
-
-    toast.success(`Example "${name}" created successfully`);
-    setCreateExampleModalOpen(false);
-  };
-
   const folderItems = sortByNameThenSequence(filter(item.items, (i) => isItemAFolder(i) && !i.isTransient));
-  const appItems = sortItemsBySequence(filter(item.items, (i) => i.type === 'app' && !i.isTransient));
-  const requestItems = sortItemsBySequence(filter(item.items, (i) => isItemARequest(i) && !i.isTransient));
-  const showEmptyFolderMessage
-    = isFolder && !hasSearchText && !folderItems?.length && !appItems?.length && !requestItems?.length;
-
+  const appItems = [...filter(item.items, (i) => i.type === 'app' && !i.isTransient)].sort((a, b) => a.seq - b.seq);
+  const requestItems = [...filter(item.items, (i) => isItemARequest(i) && !i.isTransient)].sort((a, b) => a.seq - b.seq);
+  const showEmptyFolderMessage = isFolder && !hasSearchText && !folderItems?.length && !appItems?.length && !requestItems?.length;
   const emptyFolderMenuItems = createEmptyStateMenuItems({ dispatch, collection, itemUid: item.uid });
 
-  const handleGenerateCode = () => {
-    if (
-      (item?.request?.url !== '')
-      || (item?.draft?.request?.url !== undefined && item?.draft?.request?.url !== '')
-    ) {
-      setGenerateCodeItemModalOpen(true);
-    } else {
-      toast.error('URL is required');
-    }
-  };
-
-  const viewFolderSettings = () => {
-    if (isItemAFolder(item)) {
-      if (isTabForItemPresent) {
-        dispatch(focusTab({ uid: tabUidForItem || item.uid }));
-        return;
-      }
-      dispatch(
-        addTab({
-          uid: item.uid,
-          collectionUid,
-          type: 'folder-settings',
-          pathname: item.pathname
-        })
-      );
-    }
-  };
-
-  const handleCopyItem = () => {
-    dispatch(copyRequest(item));
-    toast.success(`${getItemTypeLabel(item)} copied`);
-  };
-
-  // One-click clone: display name becomes "<source> copy"; the filesystem name
-  // uniqueness is resolved silently by electron.
-  const handleCloneItem = () => {
-    if (!isCloneable) return;
-    dispatch(cloneItem(`${item.name} copy`, sanitizeName(`${item.name} copy`), item.uid, collectionUid))
-      .then(() => toast.success(`${getItemTypeLabel(item)} cloned!`))
-      .catch((err) => toast.error(formatIpcError(err) || `An error occurred while cloning the ${getItemTypeLabel(item).toLowerCase()}`));
-  };
-
-  const handlePasteItem = () => {
-    // Determine target folder: if item is a folder, paste into it; otherwise paste into parent folder
-    let targetFolderUid = item.uid;
-    if (!isFolder) {
-      const parentFolder = findParentItemInCollection(collection, item.uid);
-      targetFolderUid = parentFolder ? parentFolder.uid : null;
-    }
-
-    dispatch(pasteItem(collectionUid, targetFolderUid))
-      .then(() => {
-        toast.success('Item pasted successfully');
-      })
-      .catch((err) => {
-        toast.error(formatIpcError(err) || 'An error occurred while pasting the item');
-      });
-  };
-
-  const handleFocus = () => {
-    setIsKeyboardFocused(true);
-    // For folders, set the folder path; for requests, set empty string (no terminal)
-    dispatch(setFocusedSidebarPath(isFolder ? item.pathname : ''));
-  };
-
-  const handleBlur = () => {
-    setIsKeyboardFocused(false);
-    dispatch(setFocusedSidebarPath(null));
-  };
-
   return (
-    <StyledWrapper className={className}>
-      {renameItemModalOpen && (
-        <RenameCollectionItem item={item} collectionUid={collectionUid} onClose={() => setRenameItemModalOpen(false)} />
-      )}
-      {deleteItemModalOpen && (
-        <DeleteCollectionItems
-          entries={[{ type: isItemAFolder(item) ? 'folder' : 'request', item, uid: item.uid, collectionUid }]}
-          onClose={() => setDeleteItemModalOpen(false)}
-        />
-      )}
-      {ignoreItemModalOpen && (
-        <IgnoreCollectionItem item={item} collectionUid={collectionUid} onClose={() => setIgnoreItemModalOpen(false)} />
-      )}
-      {newRequestModalOpen && (
-        <NewRequest item={item} collectionUid={collectionUid} onClose={() => setNewRequestModalOpen(false)} />
-      )}
-      {newFolderModalOpen && (
-        <NewFolder item={item} collectionUid={collectionUid} onClose={() => setNewFolderModalOpen(false)} />
-      )}
-      {newAppModalOpen && (
-        <NewApp item={item} collectionUid={collectionUid} onClose={() => setNewAppModalOpen(false)} />
-      )}
-      {runCollectionModalOpen && (
-        <RunCollectionItem collectionUid={collectionUid} item={item} onClose={() => setRunCollectionModalOpen(false)} />
-      )}
-      {generateCodeItemModalOpen && (
-        <GenerateCodeItem collectionUid={collectionUid} item={item} onClose={() => setGenerateCodeItemModalOpen(false)} />
-      )}
-      {itemInfoModalOpen && (
-        <CollectionItemInfo item={item} onClose={() => setItemInfoModalOpen(false)} />
-      )}
-      <CreateExampleModal
-        isOpen={createExampleModalOpen}
-        onClose={() => setCreateExampleModalOpen(false)}
-        onSave={handleCreateExample}
-        title="Create Response Example"
-        initialName={getInitialExampleName(item)}
-      />
-      <div
-        className={itemRowClassName}
-        ref={ref}
-        tabIndex={0}
-        onFocus={handleFocus}
-        onBlur={handleBlur}
-        onContextMenu={handleContextMenu}
-        data-testid="sidebar-collection-item-row"
-        data-selected={isSelected ? 'true' : undefined}
-      >
-        <div className="flex items-center h-full w-full">
-          {indents && indents.length
-            ? indents.map((i) => (
-                <div
-                  onClick={handleClick}
-                  onDoubleClick={handleDoubleClick}
-                  className="indent-block"
-                  key={i}
-                  style={{ width: 16, minWidth: 16, height: '100%' }}
-                >
-                &nbsp;{/* Indent */}
-                </div>
-              ))
-            : null}
-          <div
-            className="flex flex-grow items-center h-full overflow-hidden"
-            style={{ paddingLeft: 8 }}
-            onClick={handleClick}
-            onDoubleClick={handleDoubleClick}
-          >
-
-            {isFolder ? (
-              <ActionIcon style={{ width: 16, minWidth: 16 }}>
-                <IconChevronRight
-                  size={16}
-                  strokeWidth={2}
-                  className={iconClassName}
-                  style={{ color: 'rgb(160 160 160)' }}
-                  onClick={handleFolderCollapse}
-                  onDoubleClick={handleFolderDoubleClick}
-                  data-testid="folder-chevron"
-                />
-              </ActionIcon>
-            ) : hasExamples ? (
-              <ActionIcon style={{ width: 16, minWidth: 16 }}>
-                <IconChevronRight
-                  size={16}
-                  strokeWidth={2}
-                  className={examplesIconClassName}
-                  style={{ color: 'rgb(160 160 160)' }}
-                  onClick={handleExamplesCollapse}
-                  onDoubleClick={handleExamplesDoubleClick}
-                  data-testid="request-item-chevron"
-                />
-              </ActionIcon>
-            ) : null}
-
-            <div className="ml-1 flex w-full h-full items-center overflow-hidden">
-              <CollectionItemIcon item={item} />
-              <span className="item-name" title={item.name}>
-                {item.name}
-              </span>
-            </div>
-          </div>
-          {!isDragging && !isMultiSelected && (
-            <div className="pr-2 collection-actions">
-              <MenuDropdown
-                ref={menuDropdownRef}
-                items={buildMenuItems()}
-                placement="bottom-start"
-                data-testid="collection-item-menu"
-                popperOptions={{ strategy: 'fixed' }}
-                appendTo={dropdownContainerRef?.current || document.body}
-              >
-                <ActionIcon className="menu-icon">
-                  <IconDots size={18} className="collection-item-menu-icon" />
-                </ActionIcon>
-              </MenuDropdown>
-            </div>
-          )}
-        </div>
-      </div>
+    <CollectionItemRow
+      item={item}
+      collectionUid={collectionUid}
+      collectionPathname={collectionPathname}
+      searchText={searchText}
+      openBulkMenu={openBulkMenu}
+    >
       {!itemIsCollapsed ? (
         <div>
-          {folderItems && folderItems.length
-            ? folderItems.map((i) => {
-                return <CollectionItem key={i.uid} item={i} collectionUid={collectionUid} collectionPathname={collectionPathname} searchText={searchText} openBulkMenu={openBulkMenu} />;
-              })
-            : null}
-          {appItems && appItems.length
-            ? appItems.map((i) => {
-                return <CollectionItem key={i.uid} item={i} collectionUid={collectionUid} collectionPathname={collectionPathname} searchText={searchText} openBulkMenu={openBulkMenu} />;
-              })
-            : null}
-          {requestItems && requestItems.length
-            ? requestItems.map((i) => {
-                return <CollectionItem key={i.uid} item={i} collectionUid={collectionUid} collectionPathname={collectionPathname} searchText={searchText} openBulkMenu={openBulkMenu} />;
-              })
-            : null}
+          {folderItems.map((i) => (
+            <CollectionItem key={i.uid} item={i} collectionUid={collectionUid} collectionPathname={collectionPathname} searchText={searchText} openBulkMenu={openBulkMenu} />
+          ))}
+          {appItems.map((i) => (
+            <CollectionItem key={i.uid} item={i} collectionUid={collectionUid} collectionPathname={collectionPathname} searchText={searchText} openBulkMenu={openBulkMenu} />
+          ))}
+          {requestItems.map((i) => (
+            <CollectionItem key={i.uid} item={i} collectionUid={collectionUid} collectionPathname={collectionPathname} searchText={searchText} openBulkMenu={openBulkMenu} />
+          ))}
           {showEmptyFolderMessage ? (
             <div className="empty-folder-message">
               {range(item.depth + 1).map((i) => (
@@ -888,24 +71,7 @@ const CollectionItem = ({ item, collectionUid, collectionPathname, searchText, o
           ) : null}
         </div>
       ) : null}
-
-      {/* Show examples when expanded (only for HTTP requests) */}
-      {isItemARequest(item) && item.type === 'http-request' && examplesExpanded && hasExamples && (
-        <div>
-          {(item.examples || []).map((example, index) => {
-            return (
-              <ExampleItem
-                key={example.uid || index}
-                example={example}
-                item={item}
-                index={index}
-                collection={collection}
-              />
-            );
-          })}
-        </div>
-      )}
-    </StyledWrapper>
+    </CollectionItemRow>
   );
 };
