+++ packages/bruno-app/src/components/Sidebar/Collections/Collection/CollectionRow/index.jsx
@@ -1,0 +1,644 @@
+import React, { useState, useRef, useEffect, useMemo } from 'react';
+import classnames from 'classnames';
+import { uuid } from 'utils/common';
+import { useDrop, useDrag } from 'react-dnd';
+import { getEmptyImage } from 'react-dnd-html5-backend';
+import {
+  IconChevronRight,
+  IconDots,
+  IconLoader2,
+  IconFilePlus,
+  IconFolderPlus,
+  IconCopy,
+  IconClipboard,
+  IconPlayerPlay,
+  IconEdit,
+  IconShare,
+  IconFoldDown,
+  IconX,
+  IconSettings,
+  IconTerminal2,
+  IconFolder,
+  IconBook,
+  IconServer,
+  IconFileArrowRight,
+  IconAppWindow
+} from '@tabler/icons';
+import OpenAPISyncIcon from 'components/Icons/OpenAPISync';
+import { toggleCollection, collapseFullCollection, clearSidebarSelection } from 'providers/ReduxStore/slices/collections';
+import { mountCollection, moveCollectionAndPersist, handleMultipleCollectionItemsDrop, pasteItem, showInFolder, saveCollectionSecurityConfig } from 'providers/ReduxStore/slices/collections/actions';
+import { useDispatch, useSelector } from 'react-redux';
+import { addTab, makeTabPermanent } from 'providers/ReduxStore/slices/tabs';
+import { setFocusedSidebarPath } from 'providers/ReduxStore/slices/app';
+import toast from 'react-hot-toast';
+import NewRequest from 'components/Sidebar/NewRequest';
+import NewFolder from 'components/Sidebar/NewFolder';
+import NewApp from 'components/Sidebar/NewApp';
+import RemoveCollections from '../RemoveCollections';
+import MoveToWorkspace from '../MoveToWorkspace';
+import { isPathExternalToBasePath } from 'utils/common/path';
+import { doesCollectionHaveItemsMatchingSearchText } from 'utils/collections/search';
+import { getSortedDraggedItems, getSelectionInfo } from 'utils/collections';
+import { isTabForItemActive } from 'src/selectors/tab';
+
+import RenameCollection from '../RenameCollection';
+import StyledWrapper from '../StyledWrapper';
+import CloneCollection from '../CloneCollection';
+import { scrollToTheActiveTab } from 'utils/tabs';
+import ShareCollection from 'components/ShareCollection/index';
+import GenerateDocumentation from '../GenerateDocumentation';
+import { getRevealInFolderLabel } from 'utils/common/platform';
+import { openDevtoolsAndSwitchToTerminal } from 'utils/terminal';
+import ActionIcon from 'ui/ActionIcon';
+import MenuDropdown from 'ui/MenuDropdown';
+import { useSidebarAccordion } from 'components/Sidebar/SidebarAccordionContext';
+import useKeybinding from 'hooks/useKeybinding';
+import { useBetaFeature, BETA_FEATURES } from 'utils/beta-features';
+import StatusBadge from 'ui/StatusBadge';
+import CreateMockServerModal from 'components/MockServer/CreateMockServerModal';
+import useSidebarSelectionClick from 'hooks/useSidebarSelectionClick';
+import useMultiSelectDragDisabled from 'hooks/useMultiSelectDragDisabled';
+
+const CollectionRow = ({ collection, searchText, openBulkMenu, children }) => {
+  const isMockServerEnabled = useBetaFeature(BETA_FEATURES.MOCK_SERVER);
+  const { dropdownContainerRef } = useSidebarAccordion();
+  const [showNewFolderModal, setShowNewFolderModal] = useState(false);
+  const [showNewRequestModal, setShowNewRequestModal] = useState(false);
+  const [showNewAppModal, setShowNewAppModal] = useState(false);
+  const [showRenameCollectionModal, setShowRenameCollectionModal] = useState(false);
+  const [showCloneCollectionModalOpen, setShowCloneCollectionModalOpen] = useState(false);
+  const [showShareCollectionModal, setShowShareCollectionModal] = useState(false);
+  const [showGenerateDocumentationModal, setShowGenerateDocumentationModal] = useState(false);
+  const [showRemoveCollectionModal, setShowRemoveCollectionModal] = useState(false);
+  const [showMoveToWorkspaceModal, setShowMoveToWorkspaceModal] = useState(false);
+  const [showCreateMockServerModal, setShowCreateMockServerModal] = useState(false);
+  const [dropType, setDropType] = useState(null);
+  const [isKeyboardFocused, setIsKeyboardFocused] = useState(false);
+  const dispatch = useDispatch();
+  const isLoading = collection.isLoading;
+  const collectionRef = useRef(null);
+
+  const isCollectionFocused = useSelector(isTabForItemActive({ itemUid: collection.uid }));
+  const { hasCopiedItems } = useSelector((state) => state.app.clipboard);
+  const selectedSidebarUids = useSelector((state) => state.collections.selectedSidebarUids);
+  const isSelected = selectedSidebarUids.includes(collection.uid);
+  const isMultiSelected = isSelected && selectedSidebarUids.length > 1;
+  const handleSelectionClick = useSidebarSelectionClick({ uid: collection.uid, searchText });
+  const menuDropdownRef = useRef(null);
+
+  // 'Move into Workspace' is available for collections opened from outside the current workspace.
+  const activeWorkspace = useSelector((state) =>
+    state.workspaces.workspaces.find((w) => w.uid === state.workspaces.activeWorkspaceUid)
+  );
+  const workspaces = useSelector((state) => state.workspaces.workspaces);
+  const collectionSortOrder = useSelector((state) => state.collections.collectionSortOrder);
+  const allCollections = useSelector((state) => state.collections.collections);
+  const isMoveToWorkspaceVisible = isPathExternalToBasePath(activeWorkspace?.pathname, collection.pathname);
+
+  const isDragDisabled = useMultiSelectDragDisabled({ isSelected, selectedSidebarUids, allCollections });
+
+  // When dragging a multi-selected collection, carry all other selected collections along
+  // so dropping one reorders the entire selection together. Mixed selections (a collection
+  // alongside a folder/request) are drag-disabled entirely, so this only ever needs to
+  // handle collection-only selections.
+  const multiDragItems = useMemo(() => {
+    if (!isSelected || !selectedSidebarUids || selectedSidebarUids.length < 2) return null;
+    const { effectiveSelection, hasFolder, hasRequest } = getSelectionInfo({ collections: allCollections, selectedUids: selectedSidebarUids });
+    if (hasFolder || hasRequest) return null;
+    const collectionEntries = effectiveSelection.filter((entry) => entry.type === 'collection');
+    return collectionEntries.map((entry) => entry.collection);
+  }, [isSelected, selectedSidebarUids, allCollections]);
+
+  // Open the OpenAPI Sync tab
+  const openOpenAPISyncTab = () => {
+    ensureCollectionIsMounted();
+    dispatch(
+      addTab({
+        uid: uuid(),
+        collectionUid: collection.uid,
+        type: 'openapi-sync'
+      })
+    );
+  };
+
+  const openMockServerDashboard = () => {
+    ensureCollectionIsMounted();
+    setShowCreateMockServerModal(true);
+  };
+
+  const handleRun = () => {
+    dispatch(
+      addTab({
+        uid: uuid(),
+        collectionUid: collection.uid,
+        type: 'collection-runner'
+      })
+    );
+  };
+
+  const ensureCollectionIsMounted = () => {
+    if (collection.mountStatus === 'mounted' || collection.mountStatus === 'mounting') {
+      return;
+    }
+    dispatch(mountCollection({
+      collectionUid: collection.uid,
+      collectionPathname: collection.pathname,
+      brunoConfig: collection.brunoConfig
+    }));
+  };
+
+  const hasSearchText = searchText && searchText?.trim()?.length;
+  const collectionIsCollapsed = hasSearchText ? false : collection.collapsed;
+
+  const iconClassName = classnames({
+    'rotate-90': !collectionIsCollapsed
+  });
+
+  const handleClick = (event) => {
+    if (handleSelectionClick(event)) return;
+    if (event.detail != 1) return;
+
+    // Check if the click came from the chevron icon
+    const isChevronClick = event.target.closest('svg')?.classList.contains('chevron-icon');
+
+    setTimeout(scrollToTheActiveTab, 50);
+
+    ensureCollectionIsMounted();
+
+    if (collection.collapsed) {
+      dispatch(toggleCollection(collection.uid));
+      // Set default jsSandboxMode to 'safe' if not present and save to disk
+      if (!collection.securityConfig?.jsSandboxMode) {
+        dispatch(saveCollectionSecurityConfig(collection.uid, {
+          jsSandboxMode: 'safe'
+        }));
+      }
+    }
+
+    if (!isChevronClick) {
+      dispatch(
+        addTab({
+          uid: collection.uid,
+          collectionUid: collection.uid,
+          type: 'collection-settings'
+        })
+      );
+    }
+  };
+
+  const handleDoubleClick = (_event) => {
+    dispatch(makeTabPermanent({ uid: collection.uid }));
+  };
+
+  const handleCollectionCollapse = (e) => {
+    e.stopPropagation();
+    e.preventDefault();
+    ensureCollectionIsMounted();
+    dispatch(toggleCollection(collection.uid));
+  };
+
+  // prevent the parent's double-click handler from firing
+  const handleCollectionDoubleClick = (e) => {
+    e.stopPropagation();
+    e.preventDefault();
+  };
+
+  const handleRightClick = (event) => {
+    event.preventDefault();
+    event.stopPropagation();
+
+    if (isMultiSelected) {
+      openBulkMenu(event);
+      return;
+    }
+
+    // Otherwise, show the regular menu dropdown
+    const _menuDropdown = menuDropdownRef.current;
+    if (_menuDropdown) {
+      _menuDropdown.toggle();
+    }
+  };
+
+  const handleCollapseFullCollection = () => {
+    dispatch(collapseFullCollection({ collectionUid: collection.uid }));
+  };
+
+  const viewCollectionSettings = () => {
+    dispatch(
+      addTab({
+        uid: collection.uid,
+        collectionUid: collection.uid,
+        type: 'collection-settings'
+      })
+    );
+  };
+
+  const handleShowInFolder = () => {
+    dispatch(showInFolder(collection.pathname)).catch((error) => {
+      console.error('Error opening the folder', error);
+      toast.error('Error opening the folder');
+    });
+  };
+
+  const handlePasteItem = () => {
+    dispatch(pasteItem(collection.uid, null))
+      .then(() => {
+        toast.success('Item pasted successfully');
+      })
+      .catch((err) => {
+        toast.error(err ? err.message : 'An error occurred while pasting the item');
+      });
+  };
+
+  // Sidebar shortcuts — only active when this collection has keyboard focus
+  useKeybinding('cloneItem', () => {
+    setShowCloneCollectionModalOpen(true);
+    return false;
+  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
+
+  useKeybinding('renameItem', () => {
+    setShowRenameCollectionModal(true);
+    return false;
+  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
+
+  useKeybinding('pasteItem', () => {
+    handlePasteItem();
+    return false;
+  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
+
+  useKeybinding('newRequest', () => {
+    setShowNewRequestModal(true);
+    return false;
+  }, { enabled: isKeyboardFocused, deps: [isKeyboardFocused] });
+
+  const handleFocus = () => {
+    setIsKeyboardFocused(true);
+    dispatch(setFocusedSidebarPath(collection.pathname));
+  };
+
+  const handleBlur = () => {
+    setIsKeyboardFocused(false);
+    dispatch(setFocusedSidebarPath(null));
+  };
+
+  const isCollectionItem = (itemType) => {
+    return itemType === 'collection-item';
+  };
+
+  const [{ isDragging }, drag, dragPreview] = useDrag({
+    type: isDragDisabled ? 'disabled-drag' : 'collection',
+    item: {
+      ...collection,
+      wasSelected: isSelected,
+      ...(multiDragItems ? { multiSelectedItems: multiDragItems } : {})
+    },
+    collect: (monitor) => ({
+      isDragging: monitor.isDragging()
+    }),
+    options: {
+      dropEffect: 'move'
+    }
+  });
+
+  const [{ isOver }, drop] = useDrop({
+    accept: ['collection', 'collection-item'],
+    hover: (_draggedItem, monitor) => {
+      const itemType = monitor.getItemType();
+      if (isCollectionItem(itemType)) {
+        // For collection items, always show full highlight (inside drop)
+        setDropType('inside');
+      } else {
+        // For collections, show line indicator (above drop)
+        setDropType('above');
+      }
+    },
+    drop: async (draggedItem, monitor) => {
+      const itemType = monitor.getItemType();
+      if (isCollectionItem(itemType)) {
+        // Lazy-unmounted workspace collections are droppable in the sidebar but
+        // have no watcher yet — mount first so the move writes through and the UI updates.
+        if (collection.mountStatus !== 'mounted' && collection.mountStatus !== 'mounting') {
+          await dispatch(mountCollection({
+            collectionUid: collection.uid,
+            collectionPathname: collection.pathname,
+            brunoConfig: collection.brunoConfig
+          }));
+        }
+
+        const draggedItems = getSortedDraggedItems({
+          draggedItem,
+          allCollections,
+          workspaces,
+          activeWorkspace,
+          collectionSortOrder,
+          searchText
+        });
+
+        const validDraggedItems = draggedItems.filter((dragged) => dragged.uid !== collection.uid);
+
+        if (validDraggedItems.length > 0) {
+          dispatch(handleMultipleCollectionItemsDrop({ targetItem: collection, draggedItems: validDraggedItems, dropType: 'inside', collectionUid: collection.uid }));
+        }
+
+        if (draggedItem.wasSelected) {
+          dispatch(clearSidebarSelection());
+        }
+      } else {
+        const draggedItems = getSortedDraggedItems({
+          draggedItem,
+          allCollections,
+          workspaces,
+          activeWorkspace,
+          collectionSortOrder,
+          searchText
+        });
+
+        const validDraggedItems = draggedItems.filter((dragged) => dragged.uid !== collection.uid);
+
+        for (const dragged of validDraggedItems) {
+          await dispatch(moveCollectionAndPersist({ draggedItem: dragged, targetItem: collection }));
+        }
+
+        if (draggedItem.wasSelected) {
+          dispatch(clearSidebarSelection());
+        }
+      }
+      setDropType(null);
+    },
+    canDrop: (draggedItem) => {
+      if (draggedItem.uid === collection.uid) return false;
+      return !draggedItem.multiSelectedItems?.some((i) => i.uid === collection.uid);
+    },
+    collect: (monitor) => ({
+      isOver: monitor.isOver()
+    })
+  });
+
+  drag(drop(collectionRef));
+  dragPreview(getEmptyImage(), { captureDraggingState: true });
+
+  useEffect(() => {
+    if (isCollectionFocused && collectionRef.current) {
+      try {
+        collectionRef.current.scrollIntoView({ behavior: 'smooth', block: 'nearest' });
+      } catch (err) {
+        // ignore scroll errors
+      }
+    }
+  }, [isCollectionFocused]);
+
+  // Debounce showing empty state to prevent flicker
+  // Race condition: isLoading can become false before items batch arrives from IPC
+
+  if (searchText && searchText.length) {
+    if (!doesCollectionHaveItemsMatchingSearchText(collection, searchText)) {
+      return null;
+    }
+  }
+
+  const collectionRowClassName = classnames(
+    'flex py-1 collection-name items-center relative',
+    {
+      'item-hovered': isOver && dropType === 'above', // For collection-to-collection moves (show line)
+      'drop-target': isOver && dropType === 'inside', // For collection-item drops (highlight full area)
+      'collection-focused-in-tab': isCollectionFocused && !isKeyboardFocused,
+      'collection-keyboard-focused': isKeyboardFocused,
+      'collection-selected': isSelected,
+      'drag-disabled': isDragDisabled
+    }
+  );
+
+  const menuItems = [
+    {
+      id: 'new-request',
+      leftSection: IconFilePlus,
+      label: 'New Request',
+      onClick: () => {
+        ensureCollectionIsMounted();
+        setShowNewRequestModal(true);
+      }
+    },
+    {
+      id: 'new-folder',
+      leftSection: IconFolderPlus,
+      label: 'New Folder',
+      onClick: () => {
+        ensureCollectionIsMounted();
+        setShowNewFolderModal(true);
+      }
+    },
+    {
+      id: 'new-app',
+      leftSection: IconAppWindow,
+      label: 'New App',
+      onClick: () => {
+        ensureCollectionIsMounted();
+        setShowNewAppModal(true);
+      }
+    },
+    {
+      id: 'run',
+      leftSection: IconPlayerPlay,
+      label: 'Run',
+      onClick: () => {
+        ensureCollectionIsMounted();
+        handleRun();
+      }
+    },
+    {
+      id: 'clone',
+      leftSection: IconCopy,
+      label: 'Clone',
+      testId: 'clone-collection',
+      onClick: () => {
+        setShowCloneCollectionModalOpen(true);
+      }
+    },
+    {
+      id: 'sync-openapi',
+      leftSection: OpenAPISyncIcon,
+      label: 'OpenAPI',
+      onClick: openOpenAPISyncTab
+    },
+    ...(hasCopiedItems
+      ? [
+          {
+            id: 'paste',
+            leftSection: IconClipboard,
+            label: 'Paste',
+            onClick: handlePasteItem
+          }
+        ]
+      : []),
+    {
+      id: 'rename',
+      leftSection: IconEdit,
+      label: 'Rename',
+      onClick: () => {
+        setShowRenameCollectionModal(true);
+      }
+    },
+    {
+      id: 'share',
+      leftSection: IconShare,
+      label: 'Share',
+      onClick: () => {
+        ensureCollectionIsMounted();
+        setShowShareCollectionModal(true);
+      }
+    },
+    {
+      id: 'generate-docs',
+      leftSection: IconBook,
+      label: 'Generate Docs',
+      onClick: () => {
+        ensureCollectionIsMounted();
+        setShowGenerateDocumentationModal(true);
+      }
+    },
+    {
+      id: 'collapse',
+      leftSection: IconFoldDown,
+      label: 'Collapse',
+      onClick: handleCollapseFullCollection
+    },
+    {
+      id: 'show-in-folder',
+      leftSection: IconFolder,
+      label: getRevealInFolderLabel(),
+      onClick: handleShowInFolder
+    },
+    ...(isMockServerEnabled ? [{
+      id: 'create-mock-server',
+      leftSection: IconServer,
+      label: 'Create Mock server',
+      rightSection: <StatusBadge status="info" size="xs">Beta</StatusBadge>,
+      onClick: openMockServerDashboard
+    }] : []),
+    {
+      id: 'divider-1',
+      type: 'divider'
+    },
+    {
+      id: 'settings',
+      leftSection: IconSettings,
+      label: 'Settings',
+      onClick: viewCollectionSettings
+    },
+    {
+      id: 'terminal',
+      leftSection: IconTerminal2,
+      label: 'Open in Terminal',
+      onClick: async () => {
+        const collectionCwd = collection.pathname;
+        await openDevtoolsAndSwitchToTerminal(dispatch, collectionCwd);
+      }
+    },
+    ...(isMoveToWorkspaceVisible
+      ? [
+          {
+            id: 'move-to-workspace',
+            leftSection: IconFileArrowRight,
+            label: 'Move into Workspace',
+            testId: 'move-collection-to-workspace',
+            onClick: () => {
+              setShowMoveToWorkspaceModal(true);
+            }
+          }
+        ]
+      : []),
+    {
+      id: 'remove',
+      leftSection: IconX,
+      label: 'Remove',
+      onClick: () => {
+        setShowRemoveCollectionModal(true);
+      }
+    }
+  ];
+
+  return (
+    <StyledWrapper className="flex flex-col" id={`collection-${collection.name.replace(/\s+/g, '-').toLowerCase()}`}>
+      {showNewRequestModal && <NewRequest collectionUid={collection.uid} onClose={() => setShowNewRequestModal(false)} />}
+      {showNewFolderModal && <NewFolder collectionUid={collection.uid} onClose={() => setShowNewFolderModal(false)} />}
+      {showNewAppModal && <NewApp collectionUid={collection.uid} onClose={() => setShowNewAppModal(false)} />}
+      {showRenameCollectionModal && (
+        <RenameCollection collectionUid={collection.uid} onClose={() => setShowRenameCollectionModal(false)} />
+      )}
+      {showRemoveCollectionModal && (
+        <RemoveCollections collectionUid={collection.uid} onClose={() => setShowRemoveCollectionModal(false)} />
+      )}
+      {showMoveToWorkspaceModal && (
+        <MoveToWorkspace collectionUid={collection.uid} onClose={() => setShowMoveToWorkspaceModal(false)} />
+      )}
+      {showShareCollectionModal && (
+        <ShareCollection collectionUid={collection.uid} onClose={() => setShowShareCollectionModal(false)} />
+      )}
+      {showGenerateDocumentationModal && (
+        <GenerateDocumentation collectionUid={collection.uid} onClose={() => setShowGenerateDocumentationModal(false)} />
+      )}
+      {showCloneCollectionModalOpen && (
+        <CloneCollection collectionUid={collection.uid} onClose={() => setShowCloneCollectionModalOpen(false)} />
+      )}
+      {showCreateMockServerModal && (
+        <CreateMockServerModal
+          defaultCollectionUid={collection.uid}
+          onClose={() => setShowCreateMockServerModal(false)}
+        />
+      )}
+      <div
+        className={collectionRowClassName}
+        ref={collectionRef}
+        tabIndex={0}
+        onFocus={handleFocus}
+        onBlur={handleBlur}
+        data-testid="sidebar-collection-row"
+        data-selected={isSelected ? 'true' : undefined}
+      >
+        <div
+          className="flex flex-grow items-center overflow-hidden"
+          onClick={handleClick}
+          onDoubleClick={handleDoubleClick}
+          onContextMenu={handleRightClick}
+        >
+          <ActionIcon style={{ width: 16, minWidth: 16 }}>
+            <IconChevronRight
+              size={16}
+              strokeWidth={2}
+              className={`chevron-icon ${iconClassName}`}
+              style={{ width: 16, minWidth: 16, color: 'rgb(160 160 160)' }}
+              data-testid="collection-chevron"
+              onClick={handleCollectionCollapse}
+              onDoubleClick={handleCollectionDoubleClick}
+            />
+          </ActionIcon>
+          <div className="ml-1 w-full" id="sidebar-collection-name" title={collection.name}>
+            {collection.name}
+          </div>
+          {isLoading ? <IconLoader2 className="animate-spin mx-1" size={18} strokeWidth={1.5} /> : null}
+        </div>
+        {!isDragging && !isMultiSelected && (
+          <div>
+            <div className="pr-2">
+              <MenuDropdown
+                ref={menuDropdownRef}
+                items={menuItems}
+                placement="bottom-start"
+                appendTo={dropdownContainerRef?.current || document.body}
+                popperOptions={{ strategy: 'fixed' }}
+                data-testid="collection-actions"
+              >
+                <ActionIcon className="collection-actions">
+                  <IconDots size={18} />
+                </ActionIcon>
+              </MenuDropdown>
+            </div>
+          </div>
+        )}
+      </div>
+      {children}
+    </StyledWrapper>
+  );
+};
+
+export default CollectionRow;
